+++ nav2_core/include/nav2_core/planner_exceptions.hpp
