+++ nav2_dwb_controller/dwb_core/src/dwb_local_planner.cpp
@@ -253,7 +253,7 @@ DWBLocalPlanner::computeVelocityCommands(
     geometry_msgs::msg::TwistStamped cmd_vel;
     cmd_vel.twist = nav_2d_utils::twist2Dto3D(cmd_vel2d.velocity);
     return cmd_vel;
-  } catch (const nav2_core::PlannerException & e) {
+  } catch (const nav2_core::ControllerException & e) {
     pub_->publishEvaluation(results);
     throw;
   }
@@ -441,7 +441,7 @@ DWBLocalPlanner::transformGlobalPlan(
   const nav_2d_msgs::msg::Pose2DStamped & pose)
 {
   if (global_plan_.poses.empty()) {
-    throw nav2_core::PlannerException("Received plan with zero length");
+    throw nav2_core::ControllerException("Received plan with zero length");
   }
 
   // let's get the pose of the robot in the frame of the plan
@@ -451,7 +451,7 @@ DWBLocalPlanner::transformGlobalPlan(
       robot_pose, transform_tolerance_))
   {
     throw dwb_core::
-          PlannerTFException("Unable to transform robot pose into global plan's frame");
+          ControllerTFException("Unable to transform robot pose into global plan's frame");
   }
 
   // we'll discard points on the plan that are outside the local costmap
@@ -535,7 +535,7 @@ DWBLocalPlanner::transformGlobalPlan(
   }
 
   if (transformed_plan.poses.empty()) {
-    throw nav2_core::PlannerException("Resulting plan has 0 poses in it.");
+    throw nav2_core::ControllerException("Resulting plan has 0 poses in it.");
   }
   return transformed_plan;
 }
