+++ nav2_dwb_controller/dwb_core/src/trajectory_utils.cpp
@@ -49,7 +49,7 @@ const geometry_msgs::msg::Pose2D & getClosestPose(
   rclcpp::Duration goal_time = rclcpp::Duration::from_seconds(time_offset);
   const unsigned int num_poses = trajectory.poses.size();
   if (num_poses == 0) {
-    throw nav2_core::PlannerException("Cannot call getClosestPose on empty trajectory.");
+    throw nav2_core::ControllerException("Cannot call getClosestPose on empty trajectory.");
   }
   unsigned int closest_index = num_poses;
   double closest_diff = 0.0;
@@ -73,7 +73,7 @@ geometry_msgs::msg::Pose2D projectPose(
   rclcpp::Duration goal_time = rclcpp::Duration::from_seconds(time_offset);
   const unsigned int num_poses = trajectory.poses.size();
   if (num_poses == 0) {
-    throw nav2_core::PlannerException("Cannot call projectPose on empty trajectory.");
+    throw nav2_core::ControllerException("Cannot call projectPose on empty trajectory.");
   }
   if (goal_time <= (trajectory.time_offsets[0])) {
     return trajectory.poses[0];
